rm_kernel: Check wrong usage case

+++ scripts/set_kernel/rm_kernel.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+if [ $# -ne 1 ];
+then
+	echo "Usage: $0 <target kernel version>"
+	exit 1
+fi
+
 rm /boot/vmlinuz-$1
 rm /boot/initrd.img-$1
 rm /boot/System.map-$1
